+++ src/pycofbuilder/exceptions.py
@@ -9,7 +9,7 @@ This file implements custom exceptions for the pycofbuilder package.
 
 class BondLenghError(Exception):
     """Exception raised when the bond length is shorter than 0.8 angstrom."""
-    def __init__(self, atom_1=None, atom_2=None, dist=0, threshold=0.8):
+    def __init__(self, atom_1=None, atom_2=None, dist: int = 0.0, threshold: float = 0.8):
         self.message = 'WARNING: Atoms {} and {} are closer than {} A, {}'.format(atom_1,
                                                                                   atom_2,
                                                                                   dist,